# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "description": "Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "",
      "description": " Main server"
    }
  ],
  "paths": {
    "/admin": {
      "get": {
        "summary": "Checks if the server is alive",
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully retrieved Admin details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "summary": "admin login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully retrieved Admin details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },

    "/admin/{id}": {
      "get": {
        "summary": "get admin detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            }
          }
        ],

        "responses": {
          "200": {
            "description": "successfully retrieved Admin details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/change-password": {
      "post": {
        "summary": "change password with old-password",
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "old-password": {
                    "type": "string"
                  },
                  "new-password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reset-password-email": {
      "post": {
        "summary": "reset-password link to email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "password sent to email successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/reset-password/{id}/{token}": {
      "post": {
        "summary": "forgot password",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new-password": {
                    "type": "string"
                  },
                  "password-confirmation": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/edit": {
      "put": {
        "summary": "Edit admin profile",
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": " name of admin"
          },

          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            },
            "description": "Email of user"
          },

          {
            "in": "query",
            "name": "image",
            "schema": {
              "type": "file"
            },
            "description": "Email of user"
          }
        ],
        "responses": {
          "200": {
            "description": "admin profile succesfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "token": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "token": "Auth token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/logout": {
      "post": {
        "summary": "logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": " admin got logout",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    }
  }
}